Move Swift 3 package to just Package.swift, and Swift 5.2 package to Package@swift-5.2.swift.
